requirements_regular_env.txt
action-tutorials-py==0.9.4
ament-copyright==0.9.8
ament-cppcheck==0.9.8
ament-cpplint==0.9.8
ament-flake8==0.9.8
ament-index-python==1.1.0
ament-lint==0.9.8
ament-lint-cmake==0.9.8
ament-package==0.9.5
ament-pep257==0.9.8
ament-uncrustify==0.9.8
ament-xmllint==0.9.8
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
Cython==0.29.37
demo-nodes-py==0.9.4
domain-coordinator==0.9.2
examples-rclpy-executors==0.9.4
examples-rclpy-minimal-action-client==0.9.4
examples-rclpy-minimal-action-server==0.9.4
examples-rclpy-minimal-client==0.9.4
examples-rclpy-minimal-publisher==0.9.4
examples-rclpy-minimal-service==0.9.4
examples-rclpy-minimal-subscriber==0.9.4
filelock==3.16.1
flatbuffers==25.9.23
fonttools==4.57.0
fsspec==2025.3.0
huggingface-hub==0.30.0
humanfriendly==10.0
idna==3.11
importlib_resources==6.4.5
Jinja2==3.1.6
joint-state-publisher==2.2.0
joint-state-publisher-gui==2.2.0
kiwisolver==1.4.7
launch==0.10.10
launch-ros==0.11.7
launch-testing==0.10.10
launch-testing-ros==0.11.7
launch-xml==0.10.10
launch-yaml==0.10.10
MarkupSafe==2.1.5
matplotlib==3.7.5
mecharm==0.0.0
mecharm_communication==0.0.0
mecharm_pi==0.0.0
ml-dtypes==0.2.0
mpmath==1.3.0
myarm_300==0.0.0
myarm_c650==0.0.0
myarm_m750==0.0.0
mybuddy==0.0.0
mycobot_280==0.0.0
mycobot_280_x3pi==0.0.0
mycobot_280jn==0.0.0
mycobot_280pi==0.0.0
mycobot_320==0.0.0
mycobot_320pi==0.0.0
mycobot_600==0.0.0
mycobot_630==0.0.0
mycobot_communication==0.0.0
mycobot_description==0.0.0
mypalletizer_260==0.0.0
mypalletizer_260_pi==0.0.0
mypalletizer_communication==0.0.0
nav2_gazebo_spawner==0.3.0
networkx==3.1
numpy==1.24.4
onnx==1.17.0
onnxruntime-gpu @ file:///home/agilex/ros2_rubbishbot_ws/src/onnxruntime_gpu-1.16.0-cp38-cp38-linux_aarch64.whl#sha256=44c82c33c41a0670702c67af1f1425002d57f66d5efb971b35e86067d0e971d9
onnxslim==0.1.78
opencv-python==4.12.0.88
osrf-pycommon==0.1.11
packaging==25.0
pandas==2.0.3
pillow==10.4.0
pkg_resources==0.0.0
protobuf==5.29.5
psutil==7.1.3
py-cpuinfo==9.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
quality-of-service-demo-py==0.9.4
requests==2.32.4
ros2_trt_pose_hand==0.0.0
ros2action==0.9.13
ros2bag==0.3.11
ros2cli==0.9.13
ros2component==0.9.13
ros2doctor==0.9.13
ros2interface==0.9.13
ros2launch==0.11.7
ros2lifecycle==0.9.13
ros2multicast==0.9.13
ros2node==0.9.13
ros2param==0.9.13
ros2pkg==0.9.13
ros2run==0.9.13
ros2service==0.9.13
ros2topic==0.9.13
rosidl-runtime-py==0.9.1
rpyutils==0.2.0
rqt==1.1.2
rqt-action==1.0.2
rqt-console==1.1.2
rqt-graph==1.1.3
rqt-gui==1.1.2
rqt-gui-py==1.1.2
rqt-moveit==1.0.1
rqt-msg==1.1.1
rqt-plot==1.1.1
rqt-publisher==1.3.0
rqt-py-console==1.0.2
rqt-reconfigure==1.0.8
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==1.0.5
rqt-robot-steering==1.0.0
rqt-runtime-monitor==1.0.0
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-tf-tree==1.0.2
rqt-top==1.0.2
rqt-topic==1.3.0
rtabmap-python==0.1.0
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
sros2==0.9.5
sympy==1.13.3
teleop-twist-keyboard==2.3.2
thop==0.1.1.post2209072238
topic-monitor==0.9.4
torch @ file:///home/agilex/ros2_rubbishbot_ws/src/src/torch-2.1.0a0%2B41361538.nv23.06-cp38-cp38-linux_aarch64.whl#sha256=5112e2ef5051f1003ae2ffb545ae596377df66979d62a954d774f18009064dbe
torchvision==0.19.1
tqdm==4.67.1
turtlebot3-example==2.1.1
turtlebot3-teleop==2.1.1
typing_extensions==4.13.2
tzdata==2025.2
ultraarm==0.0.0
ultralytics==8.3.235
urllib3==2.2.3
zipp==3.20.2
